requirements.txt
# Multi-Drone Mission - Requirements
# ELKa 2026 Final Mission
# Cross-platform: Works on Linux (Arch) and Windows 10/11

# MAVLink Communication
pymavlink>=2.4.0
pyserial>=3.5  # Required for cross-platform serial port detection

# Geodetic Calculations
shapely>=2.0.0
pyproj>=3.6.0

# Computer Vision
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Utilities
pyyaml>=6.0
